chore: move slack to utils directory
