+++ pkg/utils/slack_test.go
@@ -1,4 +1,4 @@
-package kor
+package utils
 
 import (
 	"bytes"
